+++ src/mainwindow.h
@@ -52,8 +52,8 @@ struct Settings {
   void saveRecent();
   void loadRecent();
 
-  static QString highlight2string(const HighlightPatterns&hm);
-  static HighlightPatterns highlightFromString(const QString&jsonStr);
+  static QString highlight2string(const HighlightPatterns &hm);
+  static HighlightPatterns highlightFromString(const QString &jsonStr);
 
   MainWindow *_mainWindow;
   QSettings m_settings, m_highlight_settings, m_recent_files_settings, m_filters_settings;
@@ -111,6 +111,8 @@ public slots:
 
   // highlights
   void openHighlightDlg();
+  void exportHighlight();
+  void importHighlight();
 
   // filtration
   void showFltrPanelSlot();